nodejs - refactoring codegen/ to use src/api.json
