+++ wrappers/nodejs/package.json
@@ -32,6 +32,7 @@
   "devDependencies": {
     "ava": "^0.25.0",
     "cuid": "^2.1.0",
+    "json-stringify-pretty-compact": "^1.1.0",
     "standard": "*",
     "tmp": "0.0.33"
   }
