--- wrappers/nodejs/src/errorCodes.json
@@ -1,43 +1,0 @@
-{
-  "c100": "CommonInvalidParam1",
-  "c101": "CommonInvalidParam2",
-  "c102": "CommonInvalidParam3",
-  "c103": "CommonInvalidParam4",
-  "c104": "CommonInvalidParam5",
-  "c105": "CommonInvalidParam6",
-  "c106": "CommonInvalidParam7",
-  "c107": "CommonInvalidParam8",
-  "c108": "CommonInvalidParam9",
-  "c109": "CommonInvalidParam10",
-  "c110": "CommonInvalidParam11",
-  "c111": "CommonInvalidParam12",
-  "c112": "CommonInvalidState",
-  "c113": "CommonInvalidStructure",
-  "c114": "CommonIOError",
-  "c200": "WalletInvalidHandle",
-  "c201": "WalletUnknownTypeError",
-  "c202": "WalletTypeAlreadyRegisteredError",
-  "c203": "WalletAlreadyExistsError",
-  "c204": "WalletNotFoundError",
-  "c205": "WalletIncompatiblePoolError",
-  "c206": "WalletAlreadyOpenedError",
-  "c207": "WalletAccessFailed",
-  "c300": "PoolLedgerNotCreatedError",
-  "c301": "PoolLedgerInvalidPoolHandle",
-  "c302": "PoolLedgerTerminated",
-  "c303": "LedgerNoConsensusError",
-  "c304": "LedgerInvalidTransaction",
-  "c305": "LedgerSecurityError",
-  "c306": "PoolLedgerConfigAlreadyExistsError",
-  "c307": "PoolLedgerTimeout",
-  "c400": "AnoncredsRevocationRegistryFullError",
-  "c401": "AnoncredsInvalidUserRevocIndex",
-  "c402": "AnoncredsAccumulatorIsFull",
-  "c403": "AnoncredsNotIssuedError",
-  "c404": "AnoncredsMasterSecretDuplicateNameError",
-  "c405": "AnoncredsProofRejected",
-  "c406": "AnoncredsClaimRevoked",
-  "c407": "AnoncredsClaimDefAlreadyExistsError",
-  "c500": "UnknownCryptoTypeError",
-  "c600": "DidAlreadyExistsError"
-}
