+++ wrappers/nodejs/src/index.js
@@ -1,49 +1,52 @@
-var indy = require('bindings')('indy')
-var errorCodes = require('./errorCodes.json')
+var binding = require('bindings')('indy')
+var api = require('./api.json')
+var apiFunctions = require('./apiFunctions')
 
 function IndyError (err) {
   Error.call(this)
   Error.captureStackTrace(this, this.constructor)
   this.name = this.constructor.name
-  if (errorCodes['c' + err]) {
-    this.message = errorCodes['c' + err]
+  if (api.errors['c' + err]) {
+    this.message = api.errors['c' + err]
     this.indy_code = err
-    this.indy_name = errorCodes['c' + err]
+    this.indy_name = api.errors['c' + err]
   } else {
     this.message = (err + '')
   }
 }
 require('util').inherits(IndyError, Error)
 
-module.exports = {}
-module.exports.promise = {}
+var indy = {}
 
-Object.keys(indy).forEach(function (name) {
-  var fn = indy[name]
-  if (typeof fn !== 'function') {
-    return
-  }
-  module.exports[name] = function () {
+apiFunctions.forEach(function (fn) {
+  var nParams = fn.jsParams.length + 1
+
+  indy[fn.jsName] = function () {
     var args = Array.prototype.slice.call(arguments)
-    var cb = args[args.length - 1]
-    args[args.length - 1] = function (err, data) {
-      if (err) {
-        cb(new IndyError(err))
-        return
-      }
-      cb(null, data)
+    while (args.length < nParams) {
+      args.push(null)
     }
-    fn.apply(null, args)
-  }
+    args = args.slice(0, nParams)
 
-  module.exports.promise[name] = function () {
-    var args = Array.prototype.slice.call(arguments)
-    return new Promise(function (resolve, reject) {
-      args.push(function (err, data) {
-        if (err) reject(new IndyError(err))
-        else resolve(data)
+    var callback = args[nParams - 1]
+    var promise
+    if (typeof callback !== 'function') {
+      promise = new Promise(function (resolve, reject) {
+        callback = function callback (err, value) {
+          if (err) reject(err)
+          else resolve(value)
+        }
       })
-      fn.apply(null, args)
-    })
+    }
+
+    args[nParams - 1] = function (err, data) {
+      if (err) callback(new IndyError(err))
+      else callback(null, data)
+    }
+    binding[fn.jsName].apply(null, args)
+
+    return promise
   }
 })
+
+module.exports = indy
