+++ wrappers/nodejs/tests/pool.js
@@ -1,12 +1,9 @@
 var test = require('ava')
-var indy = require('../').promise
+var indy = require('../')
 var makeTestPool = require('./helpers/makeTestPool')
 
 test('pool', async function (t) {
-  var err = await t.throws(indy.create_pool_ledger_config())
-  t.is(err.message, 'Expected 3 arguments: create_pool_ledger_config(config_name, config, cb(err))')
-
-  err = await t.throws(indy.create_pool_ledger_config('', ''))
+  var err = await t.throws(indy.create_pool_ledger_config('', ''))
   t.is(err.indy_name, 'CommonInvalidParam2')
 
   err = await t.throws(indy.create_pool_ledger_config('not_a_real_pool', JSON.stringify({
